Update index.md
Commented out the Twitter logo code, as Twitter is now X, and @OpenKeychain does not exist on X.com. 

Also updated links for K-9 Mail, as it seems that it was acquired by the Thunderbird project. Also the original link for the K-9 Mail site forwards to another URL now. 

Thank you, Shalom.